Refactor Home.js and GifGrid.js
Home.js and GifGrid.js were refactored to utilize the Tenor Search
Endpoint's pos parameter to continuously fetch more gifs from the API.

This was chosen instead of the initial method of fetching a large batch
of images at once because it keeps the bandwidth usage down and a
quicker response time for the user.

This new method also allows for more gifs to be searched for since
there was a limit to how many gifs were fetched when we only made one
fetch call.