Add an overlay for each grid image on hover
The overlay will be used to show how many likes the currently has on
the gifygram app.

The images will eventually pull up a bigger overlay on top of the
entire web page to display the full image and allow give the user the
option to like the image. The likes displayed on the image overlay is
just a placeholder for now.