Implement modal system for displaying full gifs
A modal now pops open when the user clicks any gif in the gif grid on
the home page. The modal was created using the html dialog element
with a little bit of assistance from javascript to make it function
similarly to instagram's modal functionality. The ability to close the
modal when the user clicks anywhere outside of the modal contents
still needs to be implemented.

This commit also fixes the issue where it was impossible to load more
gifs if the user's viewport was big enough to contain the initial 9
gifs. This meant the user didn't have the ability to scroll which
didn't trigger the search for more gifs.